+++ src/com/ysaito/euclid/Point.java
@@ -1,23 +1,23 @@
 package com.ysaito.euclid;
 
 public abstract class Point {
-	public abstract float x();
-	public abstract float y();
+	public abstract double x();
+	public abstract double y();
 	public abstract StateId stateId();
 	
 	static public class Explicit extends Point {
-		private float mX, mY;
+		private double mX, mY;
 		private StateId.Leaf mStateId;
 		
-		public Explicit(float x, float y) {
+		public Explicit(double x, double y) {
 			mX = x;
 			mY = y;
 			mStateId = new StateId.Leaf();
 		}
 		
-		public float x() { return mX; }
-		public float y() { return mY; }
-		public void moveTo(float x, float y) {
+		public double x() { return mX; }
+		public double y() { return mY; }
+		public void moveTo(double x, double y) {
 			mX = x;
 			mY = y;
 			mStateId.updated();
@@ -29,9 +29,9 @@ public abstract class Point {
 		private int mN;
 		public final Shape shape0, shape1;
 		public StateId.NonLeaf mStateId;
-		public float mX, mY;
+		public double mX, mY;
 		
-		public Derived(Shape s0, Shape s1, float nearX, float nearY) {
+		public Derived(Shape s0, Shape s1, double nearX, double nearY) {
 			shape0 = s0;
 			shape1 = s1;
 			mStateId = new StateId.NonLeaf(s0.stateId(), s1.stateId());
@@ -46,12 +46,12 @@ public abstract class Point {
 				mY = intersection.y(mN);				
 			}
 		}
-		public float x() {
+		public double x() {
 			maybeRefreshState();
 			return mX;
 		}
 		
-		public float y() {
+		public double y() {
 			maybeRefreshState();
 			return mY;
 		}
